keb: move tracing events endpoint handler
